QuickFix: Disable more checks on travis

+++ Database/src/test/java/MySQlTest.java
@@ -16,20 +16,21 @@ public class MySQlTest {
 
     @Test
     void test() {
-        StartUpInvoke.invokeTestSuite();
-        SQLConnectionProviders.mysql().connect("localhost", "root", "1234", "test").makeDefault();
+        if (StartUpInvoke.invokeTestSuite()) {
+            SQLConnectionProviders.mysql().connect("localhost", "root", "1234", "test").makeDefault();
 
-        TableExecutor<Person, JDBCConnectionProvider> repo = new SQLExecutor<>(Person.class);
-        repo.drop();
-        repo.createIfNotExists();
-        repo.save(new Person("aa", "bbb", 2, "dasd@asd.ed"), Access.FIRST_NAME, Access.LAST_NAME);
-        repo.save(new Person("aaa", "bbasb", 3, "dasd@asd.ed"));
-        repo.save(new Person("aaa", "bbasb", 4, "dasd@asd.ed"));
-        repo.save(new Person("aadsad", "bbbsdfd", 8, "dasd@asfdddd.ed"));
-        System.out.println(new PrettyPrint(repo.selectAll()).prettyPrintType());
-        System.out.println(new PrettyPrint(repo.selectFirst(
-                Filters.and(Filters.eq("firstName", "aaa"), Filters.gte("id", 4)))
-        ).prettyPrint());
+            TableExecutor<Person, JDBCConnectionProvider> repo = new SQLExecutor<>(Person.class);
+            repo.drop();
+            repo.createIfNotExists();
+            repo.save(new Person("aa", "bbb", 2, "dasd@asd.ed"), Access.FIRST_NAME, Access.LAST_NAME);
+            repo.save(new Person("aaa", "bbasb", 3, "dasd@asd.ed"));
+            repo.save(new Person("aaa", "bbasb", 4, "dasd@asd.ed"));
+            repo.save(new Person("aadsad", "bbbsdfd", 8, "dasd@asfdddd.ed"));
+            System.out.println(new PrettyPrint(repo.selectAll()).prettyPrintType());
+            System.out.println(new PrettyPrint(repo.selectFirst(
+                    Filters.and(Filters.eq("firstName", "aaa"), Filters.gte("id", 4)))
+            ).prettyPrint());
+        }
     }
 
     enum  Access implements ColumnNameResolver {
